Refactor: MessageList 분리 시작
